+++ pkg/analytics/deploy.go
@@ -8,16 +8,6 @@ import (
 
 const deployEvent = "Deploy"
 
-type AnalyticsTracker struct {
-	TrackFn func(event string, success bool, props map[string]interface{})
-}
-
-func NewAnalyticsTracker() *AnalyticsTracker {
-	return &AnalyticsTracker{
-		TrackFn: track,
-	}
-}
-
 // DeployMetadata contains the metadata of a deploy event
 type DeployMetadata struct {
 	Success                bool
@@ -51,20 +41,4 @@ func (a *AnalyticsTracker) TrackDeploy(metadata DeployMetadata) {
 		props["error"] = metadata.Err.Error()
 	}
 	a.TrackFn(deployEvent, metadata.Success, props)
-}
-
-// DestroyMetadata contains the metadata of a destroy event
-type DestroyMetadata struct {
-	Success      bool
-	IsDestroyAll bool
-	IsRemote     bool
-}
-
-// TrackDestroy sends a tracking event to mixpanel when the user destroys a pipeline from local
-func (a *AnalyticsTracker) TrackDestroy(metadata DestroyMetadata) {
-	props := map[string]any{
-		"isDestroyAll": metadata.IsDestroyAll,
-		"isRemote":     metadata.IsRemote,
-	}
-	a.TrackFn(destroyEvent, metadata.Success, props)
 }
