+++ pkg/analytics/deploy_test.go
@@ -8,12 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-type mockEvent struct {
-	event   string
-	success bool
-	props   map[string]any
-}
-
 func TestDeployTracker(t *testing.T) {
 	tt := []struct {
 		name     string
@@ -123,80 +117,6 @@ func TestDeployTracker(t *testing.T) {
 			}
 
 			tracker.TrackDeploy(tc.metadata)
-
-			assert.Equal(t, tc.expected.event, eventReceived.event)
-			assert.Equal(t, tc.expected.success, eventReceived.success)
-			assert.Equal(t, tc.expected.props, eventReceived.props)
-		})
-	}
-}
-
-func TestDestroyTracker(t *testing.T) {
-	tt := []struct {
-		name     string
-		metadata DestroyMetadata
-		expected mockEvent
-	}{
-		{
-			name: "success destroy",
-			metadata: DestroyMetadata{
-				Success: true,
-			},
-			expected: mockEvent{
-				event:   destroyEvent,
-				success: true,
-				props: map[string]any{
-					"isDestroyAll": false,
-					"isRemote":     false,
-				},
-			},
-		},
-		{
-			name: "success destroy on remote",
-			metadata: DestroyMetadata{
-				Success:  true,
-				IsRemote: true,
-			},
-			expected: mockEvent{
-				event:   destroyEvent,
-				success: true,
-				props: map[string]any{
-					"isDestroyAll": false,
-					"isRemote":     true,
-				},
-			},
-		},
-		{
-			name: "fail destroy all",
-			metadata: DestroyMetadata{
-				Success:      false,
-				IsRemote:     false,
-				IsDestroyAll: true,
-			},
-			expected: mockEvent{
-				event:   destroyEvent,
-				success: false,
-				props: map[string]any{
-					"isDestroyAll": true,
-					"isRemote":     false,
-				},
-			},
-		},
-	}
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			eventReceived := &mockEvent{}
-			tracker := AnalyticsTracker{
-				TrackFn: func(event string, success bool, props map[string]any) {
-					eventReceived = &mockEvent{
-						event:   event,
-						success: success,
-						props:   props,
-					}
-				},
-			}
-
-			tracker.TrackDestroy(tc.metadata)
 
 			assert.Equal(t, tc.expected.event, eventReceived.event)
 			assert.Equal(t, tc.expected.success, eventReceived.success)
