+++ pkg/analytics/tracker_test.go
@@ -1,0 +1,7 @@
+package analytics
+
+type mockEvent struct {
+	event   string
+	success bool
+	props   map[string]any
+}
